Fix compiler warnings.

+++ ZipArchive.m
@@ -180,7 +180,7 @@
 								  NULL,0,
 								  NULL,//comment
 								  Z_DEFLATED,
-								  self.compression );
+								  (int)self.compression );
 	}
 	else
 	{
@@ -193,7 +193,7 @@
 								  NULL,0,
 								  NULL,//comment
 								  Z_DEFLATED,
-								  self.compression,
+								  (int)self.compression,
 								  0,
 								  15,
 								  8,
